+++ utils/token.go
@@ -5,33 +5,32 @@ import (
 
 	"github.com/dgrijalva/jwt-go"
 	"twreporter.org/go-api/models"
-
-	log "github.com/Sirupsen/logrus"
+	//log "github.com/Sirupsen/logrus"
 )
 
-// RetrieveToken ...
-func RetrieveToken(userID uint, privilege int, firstName string, lastName string, email string) (string, error) {
-	log.WithFields(log.Fields{
-		"ID":        userID,
-		"privilege": privilege,
-		"firstName": firstName,
-		"lastName":  lastName,
-		"email":     email,
-	}).Info("RetrieveToken")
+// ReporterJWTClaims JWT claims we used
+type ReporterJWTClaims struct {
+	UserID uint   `json:"userID"`
+	Email  string `json:"email"`
+	jwt.StandardClaims
+}
+
+// RetrieveToken - generate jwt token according to user's info
+func RetrieveToken(userID uint, email string) (string, error) {
+
+	// Create the Claims
+	claims := ReporterJWTClaims{
+		userID,
+		email,
+		jwt.StandardClaims{
+			ExpiresAt: time.Now().Add(time.Hour * Cfg.AppSettings.Expiration).Unix(),
+			Issuer:    Cfg.AppSettings.Protocol + "://" + Cfg.AppSettings.Host + ":" + Cfg.AppSettings.Port,
+			Audience:  Cfg.ConsumerSettings.Protocol + "://" + Cfg.ConsumerSettings.Host + ":" + Cfg.ConsumerSettings.Port,
+		},
+	}
 
 	// create the token
-	token := jwt.New(jwt.SigningMethodHS256)
-
-	// create a map to store our claims
-	claims := token.Claims.(jwt.MapClaims)
-
-	// set token claims
-	claims["userID"] = userID
-	claims["privilege"] = privilege
-	claims["firstName"] = firstName
-	claims["lastName"] = lastName
-	claims["email"] = email
-	claims["exp"] = time.Now().Add(time.Hour * Cfg.AppSettings.Expiration).Unix()
+	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
 
 	/* Sign the token with our secret */
 	tokenString, err := token.SignedString([]byte(Cfg.AppSettings.Token))